Move BlockEditor's initial state logic to Sidebar
