+++ react/components/EditorContainer/Sidebar/BlockEditor/hooks.ts
@@ -5,20 +5,17 @@ import { defineMessages } from 'react-intl'
 import {
   updateExtensionFromForm,
   getSchemaPropsOrContent,
-  getIframeImplementation,
-  getExtension,
 } from '../../../../utils/components'
 import { useEditorContext } from '../../../EditorContext'
 import ListContent from '../../graphql/ListContent.graphql'
 import { NEW_CONFIGURATION_ID } from '../consts'
 import { useFormMetaContext } from '../FormMetaContext'
 import { useModalContext } from '../ModalContext'
+import { getFormData } from '../utils'
 
 import { UseFormHandlers } from './typings'
 import {
-  getDefaultCondition,
   getDefaultConfiguration,
-  getFormData,
   omitUndefined,
   throttledUpdateExtensionFromForm,
 } from './utils'
@@ -37,8 +34,6 @@ const messages = defineMessages({
 export const useFormHandlers: UseFormHandlers = ({
   iframeRuntime,
   intl,
-  isSitewide,
-  query,
   saveContent,
   setState,
   showToast,
@@ -277,56 +272,6 @@ export const useFormHandlers: UseFormHandlers = ({
     ]
   )
 
-  const handleInitialStateSet = () => {
-    const treePath = editor.editTreePath || ''
-
-    const extension = getExtension(treePath, iframeRuntime.extensions)
-
-    const listContent = query.data && query.data.listContentWithSchema
-
-    const componentImplementation = getIframeImplementation(extension.component)
-
-    const contentSchema = listContent && JSON.parse(listContent.schemaJSON)
-
-    const configurations = listContent && listContent.content
-
-    const activeContentId =
-      extension.contentIds[extension.contentIds.length - 1]
-
-    const activeContent =
-      configurations &&
-      configurations.find(
-        configuration => configuration.contentId === activeContentId
-      )
-
-    const content =
-      (activeContent &&
-        activeContent.contentJSON &&
-        JSON.parse(activeContent.contentJSON)) ||
-      {}
-
-    const condition = activeContent
-      ? activeContent.condition
-      : getDefaultCondition({ iframeRuntime, isSitewide })
-
-    const label = activeContent && activeContent.label
-
-    const formData = getFormData({
-      componentImplementation,
-      content,
-      contentSchema,
-      iframeRuntime,
-    })
-
-    setState({
-      condition,
-      contentId: activeContentId,
-      content,
-      formData,
-      label,
-    })
-  }
-
   const handleConfigurationCreate = useCallback(
     () =>
       handleInactiveConfigurationOpen(
@@ -365,7 +310,6 @@ export const useFormHandlers: UseFormHandlers = ({
     handleFormChange,
     handleFormSave,
     handleInactiveConfigurationOpen,
-    handleInitialStateSet,
     handleLabelChange,
     handleListOpen,
   }
