+++ react/components/EditorContainer/Sidebar/BlockEditor/index.tsx
@@ -1,20 +1,17 @@
 import React from 'react'
 import { injectIntl } from 'react-intl'
-import { Spinner } from 'vtex.styleguide'
 
 import { useEditorContext } from '../../../EditorContext'
 import DeleteContentMutation from '../../mutations/DeleteContent'
-import { ListContentQueryResult } from '../../queries/ListContent'
-import { FormDataContainer } from '../typings'
+import { EditingState, FormDataContainer } from '../typings'
 
 import BlockConfigurationEditor from './BlockConfigurationEditor'
 import BlockConfigurationList from './BlockConfigurationList'
 import { useFormHandlers } from './hooks'
-import { EditingState, UseFormHandlersParams } from './typings'
+import { UseFormHandlersParams } from './typings'
 
-interface Props extends Omit<UseFormHandlersParams, 'setState' | 'state'> {
-  isSitewide: boolean
-  query: ListContentQueryResult
+type Props = Omit<UseFormHandlersParams, 'setState' | 'state'> & {
+  initialEditingState?: EditingState
 }
 
 export interface State extends EditingState {
@@ -23,9 +20,8 @@ export interface State extends EditingState {
 
 const BlockEditor = ({
   iframeRuntime,
+  initialEditingState,
   intl,
-  isSitewide,
-  query,
   saveContent,
   showToast,
 }: Props) => {
@@ -36,7 +32,7 @@ const BlockEditor = ({
       ...prevState,
       ...nextState,
     }),
-    { mode: 'editingActive' }
+    { ...initialEditingState, mode: 'editingActive' }
   )
 
   const editor = useEditorContext()
@@ -49,31 +45,16 @@ const BlockEditor = ({
     handleFormChange,
     handleFormSave,
     handleInactiveConfigurationOpen,
-    handleInitialStateSet,
     handleLabelChange,
     handleListOpen,
   } = useFormHandlers({
     iframeRuntime,
     intl,
-    isSitewide,
-    query,
     saveContent,
     setState,
     showToast,
     state,
   })
-
-  if (query.loading || !state.formData) {
-    if (!query.loading && JSON.stringify(editor.blockData) !== '{}') {
-      handleInitialStateSet()
-    }
-
-    return (
-      <div className="mt9 flex justify-center">
-        <Spinner />
-      </div>
-    )
-  }
 
   const editorCommonProps = {
     condition: (state
