+++ react/components/EditorContainer/Sidebar/typings.d.ts
@@ -1,3 +1,4 @@
+import { GetSchemaPropsOrContentFromRuntimeParams } from '../../../utils/components/typings'
 import { ListContentData } from '../queries/ListContent'
 
 export interface FormMetaContext {
@@ -27,10 +28,40 @@ export interface ModifiedSidebarComponent extends SidebarComponent {
   modifiedTreePath: string
 }
 
+export interface EditingState
+  extends Partial<Omit<ExtensionConfiguration, 'contentJSON'>> {
+  content?: Extension['content']
+  formData?: Extension['content']
+}
+
 // eslint-disable-next-line @typescript-eslint/prefer-interface
 export type FormDataContainer = {
   formData: object
 }
+
+interface GetInitialEditingStateParams {
+  data: ListContentData
+  editor: EditorContextType
+  iframeRuntime: RenderContext
+}
+
+export type GetInitialEditingState = (
+  params: GetInitialEditingStateParams
+) => EditingState
+
+export type GetDefaultCondition = ({
+  iframeRuntime,
+  isSitewide,
+}) => ExtensionConfiguration['condition']
+
+interface GetFormDataParams {
+  componentImplementation: GetSchemaPropsOrContentFromRuntimeParams['component']
+  content: GetSchemaPropsOrContentFromRuntimeParams['propsOrContent']
+  contentSchema: GetSchemaPropsOrContentFromRuntimeParams['contentSchema']
+  iframeRuntime: GetSchemaPropsOrContentFromRuntimeParams['runtime']
+}
+
+export type GetFormData = (params: GetFormDataParams) => Extension['content']
 
 export type UpdateEditorBlockData = (
   params: Pick<BlockData, 'id' | 'serverTreePath' | 'template'> & {
